Move ProcessFoldersXxx.cs to the service layer, using namespace to enable it
